set variable env
